+++ packages/apps/developer-portal/src/export.js
@@ -1,1 +1,17 @@
-export { default } from "./App"
+import SwaggerDocs from "./pages/SwaggerDocs"
+import Introduction from "./pages/Introduction"
+import ApiReference from "./pages/ApiReference"
+import OpenApiProvider from "./components/OpenApiProvider"
+import OpenApiConsumer from "./components/OpenApiConsumer"
+import Header from "./components/Header/Header"
+import Tiles from "./components/Tiles"
+
+export {
+  SwaggerDocs,
+  Introduction,
+  ApiReference,
+  OpenApiProvider,
+  OpenApiConsumer,
+  Header,
+  Tiles
+}