+++ packages/themes/Styrke/src/Theme/theme.Selects.js
@@ -1,0 +1,14 @@
+import { css } from "styled-components"
+import { Select, theming } from "@staccx/base"
+
+export const SelectStyling = theming.createVariants(
+  {
+    calculatorTerms: css`
+      height: 44px;
+      button:first-child {
+        height: 44px;
+      }
+    `
+  },
+  Select.themeProps.selectWrapper
+)